Yaza 264

+++ yaza/apis/ocspu.py
@@ -7,15 +7,16 @@ from werkzeug.utils import cached_property
 
 from yaza.apis import ModelWrapper, wraps
 from yaza.basemain import app
+from yaza.models import OCSPU, Aspect, DesignRegion
 from yaza.tools.color_tools import contrast_color, darker_color
-from yaza.qiniu_handler import delete_file
 from yaza.utils import do_commit
+from yaza.qiniu_handler import delete_file
 
 
 def split_pic_url(pic_url):
     http_, key = pic_url.split(".qiniudn.com/")
     bucket = http_.split("//")[-1]
-    return bucket, key
+    return bucket.encode('utf-8'), key.encode('utf-8')
 
 
 def delete_file_from_path(path):
@@ -54,6 +55,19 @@ class OCSPUWrapper(ModelWrapper):
             "hoveredComplementColor" if camel_case else "hovered_complement_color": self.hovered_complement_color,
             'rgb': self.rgb,
         }
+
+    def clone(self):
+        # 不clone文件
+        ret = do_commit(OCSPU(color=self.color, spu_id=self.spu_id, rgb=self.rgb))
+        for aspect in self.aspect_list:
+            new_aspect = do_commit(Aspect(name=aspect.name, ocspu_id=ret.id,
+                                          width=aspect.width,
+                                          height=aspect.height))
+            for dr in aspect.design_region_list:
+                do_commit(DesignRegion(aspect_id=new_aspect.id, name=dr.name,
+                                       width=dr.width,
+                                       height=dr.height))
+        return ret
 
     def delete(self):
         for aspect in self.aspect_list:
@@ -184,9 +198,9 @@ class DesignRegionWrapper(ModelWrapper):
     def delete(self):
         if self.pic_path:
             delete_file_from_path(self.pic_path)
-        if os.path.exists(self.edge_file):
+        if self.edge_file and os.path.exists(self.edge_file):
             os.unlink(self.edge_file)
-        if os.path.exists(self.control_point_file):
+        if self.control_point_file and os.path.exists(self.control_point_file):
             os.unlink(self.control_point_file)
 
         do_commit(self, "delete")
